DINOv2_TensorRT/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dinov2-trt"
version = "0.1.0"
description = "TensorRT-accelerated DINOv2 inference"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "opencv-python",
    "torch",
    "transformers",
    "onnx",
    "onnxsim",
    "onnxruntime-gpu",
    "tensorrt",
    "pycuda",
    "scikit-learn",
]

[tool.setuptools.packages.find]
where = ["src"]
